AudioManager work
I found a way to get the SFX to work from the sliders that Chris set up in the settings. I created an AudioMixer(with 2 groups, SFX and Music) and attached that to an empty AudioManager object using a script. That script has methods that can set the volume to any value. So, I went to the AudioSource components of the Player, Civilian, & PoliceOfficer, and set their output to SFX(MasterAudioMixer). Next I found the slider for the SFX in the hierarchy. In the SFX slider's OnValueChanged(), I dragged the AudioManager prefab, and selected the method, "AudioManager.SetSFXVolume". This allows the SFX slider to have control over the SFX volume in the game. (Sorry for the long explanation)